app/requirements.txt
fastapi
uvicorn
gunicorn
redis
prometheus-fastapi-instrumentator
numpy
faiss-cpu
torch